requirements.txt
click==8.1.7
colorama==0.4.6
pillow==11.0.0
python-nmap==0.7.1